Create The factors in the financial statement
This test can use for the factors we can get from the financial statement freely. Actually, in real stock market, when we get the data in statement, we may choose to buy the stock according to the data. In order to make sure all the statement is announced, we get the data at the deadline of the statements, and then check whether to buy it.